Calculate field order via VFMMatch prop when deinterlacing
For some reason VFM discards the _FieldBased prop by setting it to 0 on all frames. So we must calculate the field order some other way. This way seems to work but has not been fully tested.

For mode=0, it works fine. For mode=1 it seems to work fine on my constant-TFF clip. However, other rarer matches like b/u have not been tested.