Fix index of props if kernel changes FPS in deinterlacer
This fixes alignment of flags/props information if deinterlaced_tff/deinterlace_bff is not the same FPS as the clip.

Since `f` gets from the clip, the fps will be e.g., 30000/1001. If you use FPSDivisor=1 via QTGMC as the kernel for example, that means `deinterlaced_tff` and `deinterlaced_bff` clips are now 60000/1001. The props used via `f` would be 2 frames behind exponentially.

Instead we can normalize the `n` to be relative to the clip, then manually grab the props at that frame.